backend/requirements.txt
fastapi
uvicorn
python-multipart
pymupdf
langchain
langchain-groq
openai
faiss-cpu
python-dotenv
google-generativeai
PyPDF2
chromadb
langchain-google-genai
langchain-community
langchain
langchain-google-genai
google-generativeai==0.3.2
